Real Fixed Program
To stop the infinite loop we checked if closeParen + 1 became smaller by comparing it to currentIndex. If currentIndex continued to get bigger but closeParen + 1 went back to the start of the text file then the infinite loop was occurring. The if statement we made checks this and breaks the loop.