Error2 Fix
The error this time was if the file had  [] but no (), so the [] weren't used in a link so the code could not find a link and had an error. To fix this and any cases with brackets or parentheses being used from something other than for links we used a if to check that all these characters exist in a file and if not just return.