ErrorFix-test-file6
test-file6 was failing. This error is caused by the code taking the link from an image when it should not. The bug was that there was no code to check for an exclamation mark before a start bracket, so we included an if statement to check for that and break if it is present.