ErrorFix-test-file5
The error was occurring in test-file5. The error was that there was the right amount of brackets and parenthesis in the file but they were separated with text and spaces in between so when the code run the array list should be empty but wasn't. It had the fake link in it. To fix this we wrote an if statement to check that the that right after an end bracket there is a parenthesis. If there is not the loop breaks.