ErrorFix-test-file8
The error this time was in test-file8. This file had a stray start bracket after a link code, which was creating an infinite loop. To fix this we checked if the value of the start bracket's index ever got greater that that of end bracket. If so then break from the loop. We also found an error where if there is no exclamation point than that created an infinite loop so we only created the variable exclamation if an exclamation was in the file.